+++ webapp/packages/nextjs/scaffold.config.ts
@@ -8,7 +8,6 @@ export type ScaffoldConfig = {
   onlyLocalBurnerWallet: boolean;
   tatumV4Key: string;
   restApiBaseUrl: string;
-  nftContractAddress: `0x${string}`;
 };
 
 const scaffoldConfig = {
@@ -36,7 +35,6 @@ const scaffoldConfig = {
 
   tatumV4Key: process.env.NEXT_TATUM_V4_API_KEY || "",
   restApiBaseUrl: "http://localhost:3001/api/",
-  nftContractAddress: `0x340A5d8E111c2613D964C132E7Dc8964ACd477fA`,
 } as const satisfies ScaffoldConfig;
 
 export default scaffoldConfig;
