+++ webapp/packages/nextjs/types/app.ts
@@ -1,0 +1,6 @@
+export type OurNFTContractInfo = {
+  address: `0x${string}`;
+  name: string;
+  symbol: string;
+  tokenCounter: number;
+};
